Добавлен хук useAuthForm.js
